3d_game_engine/requirements.txt
pygame
PyOpenGL
pyrr
